Renaming + some adjustments
