+++ react-components/src/architecture/concrete/clipping/commands/ApplyClipCommand.ts
@@ -53,10 +53,10 @@ export class ApplyClipCommand extends RenderTargetCommand {
     const cropBox = this.getSelectedCropBoxDomainObject();
     if (cropBox !== undefined) {
       cropBox.setThisAsGlobalCropBox();
+      renderTarget.fitView();
     } else {
       ApplyClipCommand.setClippingPlanes(this.rootDomainObject);
     }
-    renderTarget.fitView();
     return true;
   }
