+++ react-components/src/architecture/concrete/example/ExampleDomainObject.ts
