+++ react-components/src/architecture/concrete/example/ExampleRenderStyle.ts
