+++ react-components/src/architecture/concrete/example/ExampleTool.ts
