+++ react-components/src/architecture/concrete/example/ExampleView.ts
