+++ react-components/src/architecture/concrete/example/commands/DeleteAllExamplesCommand.ts
