+++ react-components/src/architecture/concrete/terrain/TerrainDomainObject.ts
