+++ react-components/src/architecture/concrete/terrain/TerrainRenderStyle.ts
