+++ react-components/src/components/Architecture/CommandButton.tsx
@@ -60,7 +60,7 @@ export const CommandButton = ({
 
   return (
     <CogsTooltip
-      content={<LabelWithShortcut key={tooltip} label={tooltip} shortcut={shortcut} />}
+      content={<LabelWithShortcut label={tooltip} shortcut={shortcut} />}
       appendTo={document.body}
       placement={placement}>
       <Button
