+++ react-components/src/components/Architecture/OptionButton.tsx
@@ -64,7 +64,7 @@ export const OptionButton = ({
 
   return (
     <CogsTooltip
-      content={<LabelWithShortcut key={tooltip} label={tooltip} shortcut={shortcut} />}
+      content={<LabelWithShortcut label={tooltip} shortcut={shortcut} />}
       appendTo={document.body}
       placement={placement}>
       <Dropdown
